Tex: extend to support subparsers
Currently, Tex parser calls subparsers it reads unhandled identifier.

If the input is

   \foo{bar}

and Tex parser has no interest in "\\foo", it calls subparsers with
"\\foo". So one of subparsers can get a chance to capture "bar" if
the subparser has interests in "\\foo".
